+++ CHANGELOG.md
@@ -44,6 +44,8 @@ This name should be decided amongst the team before the release.
 
 [Full list of changes](https://github.com/topiary/topiary/compare/v0.7.3...HEAD)
 
+
+
 ### Changed
 - [#1032](https://github.com/topiary/topiary/pull/1032) Replaced the `topiary-core` test and benchmark dependencies on the `tree-sitter-json` and `tree-sitter-nickel` grammar crates with dynamic loading via `topiary-config`.
 - [#1277](https://github.com/topiary/topiary/pull/1277) Small internal refactoring of the core formatting logic.
@@ -73,6 +75,7 @@ This name should be decided amongst the team before the release.
 - [#1227](https://github.com/topiary/topiary/pull/1227) Add `@append_empty_input_softline` and `@prepend_empty_input_softline` captures, thanks to @BirdeeHub
 - [#1244](https://github.com/topiary/topiary/pull/1244) Add language injection support for injection languages known at query writing.
 - [#1254](https://github.com/topiary/topiary/pull/1254) Add Markdown formatting support with dynamic language injections for fenced code blocks (via the `@injection.language` capture).
+- [#1279](https://github.com/topiary/topiary/pull/1279) Added experimental formatting support for the Menhir language.
 
 ### Removed
 - [#1217](https://github.com/topiary/topiary/pull/1217) The `check` alias for the `check-grammar` subcommand, to avoid confusion with `topiary fmt --check`.